obfuscator/requirements.txt
pefile
icicle-emu